moved function protos used by both Radio/Dirac to Common.h
