+++ src/lib/Dirac/Dirac.h
@@ -1197,14 +1197,6 @@ project_procrustes(int N,double *J,double *J1);
 extern int
 project_procrustes_block(int N,complex double *J,complex double *J1);
 
-
-/* Extract only the phase of diagonal entries from solutions 
-   p: 8Nx1 solutions, orders as [(real,imag)vec(J1),(real,imag)vec(J2),...]
-   pout: 8Nx1 phases (exp(j*phase)) of solutions, after joint diagonalization of p
-   N: no. of 2x2 Jones matrices in p, having common unitary ambiguity
-   niter: no of iterations for Jacobi rotation */
-extern int
-extract_phases(double *p, double *pout, int N, int niter);
 /****************************** consensus_poly.c ****************************/
 /* build matrix with polynomial terms
   B : Npoly x Nf, each row is one basis function
@@ -1308,286 +1300,6 @@ sagefit_visibilities_admm_dual_pt_flt(double *u, double *v, double *w, double *x
 extern void
 openblas_set_num_threads(int num_threads);
 
-/****************************** myblas.c ****************************/
-/* BLAS wrappers */
-/* machine precision */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern double 
-dlamch(char CMACH);
-
-/* blas dcopy */
-/* y = x */
-/* read x values spaced by Nx (so x size> N*Nx) */
-/* write to y values spaced by Ny  (so y size > N*Ny) */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_dcopy(int N, double *x, int Nx, double *y, int Ny);
-
-/* blas scale */
-/* x = a. x */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_dscal(int N, double a, double *x);
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_sscal(int N, float a, float *x);
-
-/* x^T*y */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern double
-my_ddot(int N, double *x, double *y);
-
-/* ||x||_2 */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern double
-my_dnrm2(int N, double *x);
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern float
-my_fnrm2(int N, float *x);
-
-/* sum||x||_1 */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern double
-my_dasum(int N, double *x);
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern float
-my_fasum(int N, float *x);
-
-/* BLAS y = a.x + y */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_daxpy(int N, double *x, double a, double *y);
-
-/* BLAS y = a.x + y */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_daxpys(int N, double *x, int incx, double a, double *y, int incy);
-
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_saxpy(int N, float *x, float a, float *y);
-
-/* max |x|  index (start from 1...)*/
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_idamax(int N, double *x, int incx);
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_isamax(int N, float *x, int incx);
-
-/* min |x|  index (start from 1...)*/
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-int
-my_idamin(int N, double *x, int incx);
-
-/* BLAS DGEMM C = alpha*op(A)*op(B)+ beta*C */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_dgemm(char transa, char transb, int M, int N, int K, double alpha, double *A, int lda, double *B, int ldb, double beta, double *C, int ldc);
-
-/* BLAS DGEMV  y = alpha*op(A)*x+ beta*y : op 'T' or 'N' */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_dgemv(char trans, int M, int N, double alpha, double *A, int lda, double *x, int incx,  double beta, double *y, int incy);
-
-/* following routines used in LAPACK dirac */
-/* cholesky factorization: real symmetric */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_dpotrf(char uplo, int N, double *A, int lda);
-
-/* solve Ax=b using cholesky factorization */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_dpotrs(char uplo, int N, int nrhs, double *A, int lda, double *b, int ldb);
-
-/* solve Ax=b using QR factorization */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_dgels(char TRANS, int M, int N, int NRHS, double *A, int LDA, double *B, int LDB, double *WORK, int LWORK);
-
-/* A=U S VT, so V needs NOT to be transposed */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_dgesvd(char JOBU, char JOBVT, int M, int N, double *A, int LDA, double *S,
-   double *U, int LDU, double *VT, int LDVT, double *WORK, int LWORK);
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_zgesvd(char JOBU, char JOBVT, int M, int N, complex double *A, int LDA, double *S,
-   complex double *U, int LDU, complex double *VT, int LDVT, complex double *WORK, int LWORK, double *RWORK); 
-
-/* QR factorization QR=A, only TAU is used for Q, R stored in A*/
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_dgeqrf(int M, int N, double *A, int LDA, double *TAU, double *WORK, int LWORK);
-
-/* calculate Q using elementary reflections */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_dorgqr(int M,int  N,int  K,double *A,int  LDA,double *TAU,double *WORK,int  LWORK);
-
-/* solves a triangular system of equations Ax=b, A triangular */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_dtrtrs(char UPLO, char TRANS, char DIAG,int N,int  NRHS,double *A,int  LDA,double *B,int  LDB);
-
-
-/* blas ccopy */
-/* y = x */
-/* read x values spaced by Nx (so x size> N*Nx) */
-/* write to y values spaced by Ny  (so y size > N*Ny) */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_ccopy(int N, complex double *x, int Nx, complex double *y, int Ny);
-
-/* blas scale */
-/* x = a. x */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_cscal(int N, complex double a, complex double *x);
-
-
-/* BLAS y = a.x + y */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_caxpy(int N, complex double *x, complex double a, complex double *y);
-
-
-/* BLAS x^H*y */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern complex double
-my_cdot(int N, complex double *x, complex double *y);
-
-/* solve Ax=b using QR factorization */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern int
-my_zgels(char TRANS, int M, int N, int NRHS, complex double *A, int LDA, complex double *B, int LDB, complex double *WORK, int LWORK);
-extern int
-my_cgels(char TRANS, int M, int N, int NRHS, complex float *A, int LDA, complex float *B, int LDB, complex float *WORK, int LWORK);
-
-/* BLAS ZGEMM C = alpha*op(A)*op(B)+ beta*C */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_zgemm(char transa, char transb, int M, int N, int K, complex double alpha, complex double *A, int lda, complex double *B, int ldb, complex double beta, complex double *C, int ldc);
-
-/* ||x||_2 */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern double
-my_cnrm2(int N, complex double *x);
-
-
-/* blas fcopy */
-/* y = x */
-/* read x values spaced by Nx (so x size> N*Nx) */
-/* write to y values spaced by Ny  (so y size > N*Ny) */
-#ifdef USE_MIC
-__attribute__ ((target(MIC)))
-#endif
-extern void
-my_fcopy(int N, float *x, int Nx, float *y, int Ny);
-
-
-/* LAPACK eigen value expert routine, real symmetric  matrix */
-extern int
-my_dsyevx(char jobz, char range, char uplo, int N, double *A, int lda,
-  double vl, double vu, int il, int iu, double abstol, int M, double  *W,
-  double *Z, int ldz, double *WORK, int lwork, int *iwork, int *ifail);
-
-/* BLAS vector outer product
-   A= alpha x x^H + A
-*/
-extern void
-my_zher(char uplo, int N, double alpha, complex double *x, int incx, complex double *A, int lda);
-
-/****************************** barrier.c ****************************/
-typedef struct t_barrier_ {
-  int tcount; /* current no. of threads inside barrier */
-  int nthreads; /* the no. of threads the barrier works
-                with. This is a constant */
-  pthread_mutex_t enter_mutex;
-  pthread_mutex_t exit_mutex;
-  pthread_cond_t lastthread_cond;
-  pthread_cond_t exit_cond;
-} th_barrier;
-
-
-/* initialize barrier */
-/* N - no. of accomodated threads */
-extern void
-init_th_barrier(th_barrier *barrier, int N);
-
-/* destroy barrier */
-extern void
-destroy_th_barrier(th_barrier *barrier);
-
-/* the main operation of the barrier */
-extern void
-sync_barrier(th_barrier *barrier);
-
 /********* solver modes *********/
 #define SM_LM_LBFGS 1
 #define SM_OSLM_LBFGS 0
@@ -1674,76 +1386,6 @@ bfgsfit_visibilities_gpu(double *u, double *v, double *w, double *x, int N,
    int Nbase, int tilesz,  baseline_t *barr,  clus_source_t *carr, complex double *coh, int M, int Mt, double freq0, double fdelta, double *pp, double uvmin, int Nt, int max_lbfgs, int lbfgs_m, int gpu_threads, int solver_mode,  double mean_nu, double *res_0, double *res_1); 
 
 
-
-/* struct to keep histoty of last used GPU */
-typedef struct taskhist_{
-  int prev; /* last used GPU (by any thread) */
-  pthread_mutex_t prev_mutex; /* mutex to lock changing prev value */
-  unsigned int rseed; /* random seed used in rand_r() */
-} taskhist;
-
-/* structs for thread pool (reusable), using a barrier */
-/* slave thread data struct */
-typedef struct slave_tdata_ {
-  struct pipeline_ *pline; /* forward declaration */
-  int tid; /* 0,1 for 2 GPUs */
-} slave_tdata;
-
-/* pipeline struct */
-typedef struct pipeline_ {
-  void *data; /* all data needed by two threads */
-  int terminate; /* 1: terminate, default 0*/
-  pthread_t slave0;
-  pthread_t slave1;
-  slave_tdata *sd0; /* note recursive types */
-  slave_tdata *sd1;
-  th_barrier gate1;
-  th_barrier gate2;
-  pthread_attr_t attr;
-  taskhist *thst;
-} th_pipeline;
-
-/* pipeline state values */
-#ifndef PT_DO_NOTHING
-#define PT_DO_NOTHING 0
-#endif
-#ifndef PT_DO_AGPU
-#define PT_DO_AGPU 1 /* allocate GPU memory, attach GPU */
-#endif
-#ifndef PT_DO_DGPU
-#define PT_DO_DGPU 2 /* free GPU memory, detach GPU */
-#endif
-#ifndef PT_DO_WORK_LM /* plain LM */
-#define PT_DO_WORK_LM 3
-#endif
-#ifndef PT_DO_WORK_OSLM /* OS accel LM */
-#define PT_DO_WORK_OSLM 4
-#endif
-#ifndef PT_DO_WORK_RLM /* robust LM */
-#define PT_DO_WORK_RLM 5
-#endif
-#ifndef PT_DO_WORK_OSRLM /* robust LM, OS accel */
-#define PT_DO_WORK_OSRLM 6
-#endif
-#ifndef PT_DO_WORK_RTR /* RTR */
-#define PT_DO_WORK_RTR 7
-#endif
-#ifndef PT_DO_WORK_RRTR /* Robust RTR */
-#define PT_DO_WORK_RRTR 8
-#endif
-#ifndef PT_DO_WORK_NSD /* Nesterov's SD */
-#define PT_DO_WORK_NSD 9
-#endif
-#ifndef PT_DO_MEMRESET 
-#define PT_DO_MEMRESET 99
-#endif
-/* for BFGS pipeline */
-#ifndef PT_DO_CDERIV
-#define PT_DO_CDERIV 20
-#endif
-#ifndef PT_DO_CCOST
-#define PT_DO_CCOST 21
-#endif
 
 
 
@@ -1849,21 +1491,6 @@ sagefit_visibilities_dual_pt_one_gpu(double *u, double *v, double *w, double *x,
 extern int
 sagefit_visibilities_dual_pt_flt(double *u, double *v, double *w, double *x, int N,
    int Nbase, int tilesz,  baseline_t *barr,  clus_source_t *carr, complex double *coh, int M, int Mt, double freq0, double fdelta, double *pp, double uvmin, int Nt, int max_emiter, int max_iter, int max_lbfgs, int lbfgs_m, int gpu_threads, int linsolv,int solver_mode,  double nulow, double nuhigh, int randomize, double *mean_nu, double *res_0, double *res_1);
-
-/****************************** load_balance.c ****************************/
-/* select a GPU from 0,1..,max_gpu
-   in such a way to allow load balancing */
-/* also keep a global variableto ensure same GPU is 
-   not assigned to one process */
-#ifdef HAVE_CUDA
-extern void
-init_task_hist(taskhist *th);
-extern void
-destroy_task_hist(taskhist *th);
-
-extern int
-select_work_gpu(int max_gpu, taskhist *th);
-#endif
 
 /****************************** stationbeam.c ****************************/
 /*
