+++ src/lib/Radio/predict.c
@@ -24,7 +24,6 @@
 #include <stdlib.h>
 #include <pthread.h>
 #include "Radio.h"
-#include "Dirac.h"
 
 /******************** shapalet stuff **********************/
 /* evaluate Hermite polynomial value using recursion
