+++ src/lib/Radio/predict_withbeam_gpu.c
@@ -19,7 +19,6 @@
 
 
 
-#include "Dirac.h"
 #include "Radio.h"
 
 //#define CUDA_DEBUG
