+++ src/lib/Radio/residual.c
@@ -24,7 +24,6 @@
 #include <unistd.h>
 #include <stdlib.h>
 #include <pthread.h>
-#include "Dirac.h" /* for extract_phases() */
 #include "Radio.h"
 
 /* Jones matrix multiplication
